% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_gmm.R
\name{GenomeGMM}
\alias{GenomeGMM}
\title{Estimate Genome Size using Gaussian Mixture Model}
\usage{
GenomeGMM(data_file, n = 6, ...)
}
\arguments{
\item{data_file}{Path to the input k-mer frequency histogram file. The file should have two columns: the first column is k-mer frequency, the second column is the count of k-mers with that frequency.}

\item{n}{Number of Gaussian components (integer), usually set according to the expected ploidy or number of main peaks.}
}
\value{
No return value. The function outputs the estimated genome size in the console and saves a PDF plot in the working directory.
}
\description{
This function reads a k-mer frequency histogram file, fits a Gaussian mixture model (GMM), estimates genome size, and generates a PDF visualization.
}
\examples{
# Example usage:
data_file <-"examples/4merged.histo.txt"
GenomeGMM (data_file, n = 4)

}
